backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.2.0
aiohttp==3.9.1
httpx==0.26.0
python-dotenv==1.0.0
alembic==1.13.1
redis==5.0.1
celery==5.3.6
matplotlib>=3.8.0
